Move rpc_module tests to tests crate
