+++ proc-macros/tests/ui/correct/basic.rs
@@ -4,8 +4,7 @@ use std::net::SocketAddr;
 
 use jsonrpsee::proc_macros::rpc;
 use jsonrpsee::rpc_params;
-use jsonrpsee::core::{async_trait, traits::Client};
-use jsonrpsee::types::RpcResult;
+use jsonrpsee::core::{async_trait, traits::Client, RpcResult};
 use jsonrpsee::ws_client::*;
 use jsonrpsee::ws_server::{SubscriptionSink, WsServerBuilder};
