+++ proc-macros/tests/ui/correct/only_server.rs
@@ -1,8 +1,7 @@
 use std::net::SocketAddr;
 
 use jsonrpsee::proc_macros::rpc;
-use jsonrpsee::core::async_trait;
-use jsonrpsee::types::RpcResult;
+use jsonrpsee::core::{async_trait, RpcResult};
 use jsonrpsee::ws_server::{SubscriptionSink, WsServerBuilder};
 
 #[rpc(server)]
