+++ proc-macros/tests/ui/correct/parse_angle_brackets.rs
@@ -1,4 +1,4 @@
-use jsonrpsee::{proc_macros::rpc, types::RpcResult};
+use jsonrpsee::{proc_macros::rpc, core::RpcResult};
 
 fn main() {
 	#[rpc(server)]
