+++ proc-macros/tests/ui/incorrect/rpc/rpc_assoc_items.rs
@@ -6,7 +6,7 @@ pub trait AssociatedConst {
 	const WOO: usize;
 
 	#[method(name = "foo")]
-	async fn async_method(&self) -> jsonrpsee::types::RpcResult<u8>;
+	async fn async_method(&self) -> jsonrpsee::core::RpcResult<u8>;
 }
 
 #[rpc(client, server)]
@@ -14,7 +14,7 @@ pub trait AssociatedType {
 	type Woo;
 
 	#[method(name = "foo")]
-	async fn async_method(&self) -> jsonrpsee::types::RpcResult<u8>;
+	async fn async_method(&self) -> jsonrpsee::core::RpcResult<u8>;
 }
 
 fn main() {}
