+++ proc-macros/tests/ui/incorrect/rpc/rpc_conflicting_alias.rs
@@ -1,5 +1,5 @@
 use jsonrpsee::proc_macros::rpc;
-use jsonrpsee::types::RpcResult;
+use jsonrpsee::core::RpcResult;
 #[rpc(client, server)]
 pub trait DuplicatedAlias {
 	#[method(name = "foo", aliases = ["foo_dup", "foo_dup"])]
