+++ proc-macros/tests/ui/incorrect/rpc/rpc_deprecated_method.rs
@@ -6,8 +6,7 @@
 use std::net::SocketAddr;
 
 use jsonrpsee::proc_macros::rpc;
-use jsonrpsee::core::async_trait;
-use jsonrpsee::types::RpcResult;
+use jsonrpsee::core::{async_trait, RpcResult};
 use jsonrpsee::ws_client::*;
 use jsonrpsee::ws_server::WsServerBuilder;
