+++ proc-macros/tests/ui/incorrect/rpc/rpc_not_qualified.rs
@@ -3,7 +3,7 @@ use jsonrpsee::proc_macros::rpc;
 // Method without type marker, `#[method(…)]` or `#[subscription(…)]`.
 #[rpc(client, server)]
 pub trait NotQualified {
-	async fn async_method(&self) -> jsonrpsee::types::RpcResult<u8>;
+	async fn async_method(&self) -> jsonrpsee::core::RpcResult<u8>;
 }
 
 fn main() {}
