+++ proc-macros/tests/ui/incorrect/rpc/rpc_not_qualified.stderr
@@ -1,5 +1,5 @@
 error: Methods must have either 'method' or 'subscription' attribute
  --> $DIR/rpc_not_qualified.rs:6:2
   |
-6 |     async fn async_method(&self) -> jsonrpsee::types::RpcResult<u8>;
-  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+6 |     async fn async_method(&self) -> jsonrpsee::core::RpcResult<u8>;
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
