+++ proc-macros/tests/ui/incorrect/sub/sub_name_override.rs
@@ -1,4 +1,4 @@
-use jsonrpsee::{proc_macros::rpc, types::RpcResult};
+use jsonrpsee::{proc_macros::rpc, core::RpcResult};
 
 // Subscription method name conflict with notif override.
 #[rpc(client, server)]
