+++ tests/Cargo.toml
@@ -14,4 +14,5 @@ futures = { version = "0.3.14", default-features = false, features = ["std"] }
 jsonrpsee = { path = "../jsonrpsee", features = ["full"] }
 tokio = { version = "1.8", features = ["full"] }
 tracing = "0.1"
+serde = "1"
 serde_json = "1"
